Feat/protected routes

+++ src/views/order.views.js
@@ -23,6 +23,11 @@ const errors = {
     err.statusCode = 400;
     return err;
   })(),
+  invalidPermissions: (() => {
+    const err = Error("You do not have the right to alter this order");
+    err.statusCode = 403;
+    return err;
+  })()
 }
 
 function formatResponseToRole(rolelabel) {
@@ -42,63 +47,58 @@ function formatResponseToRole(rolelabel) {
     case "developer":
       return {};
     default:
-      // return {};
-      return { clientCode: 0, restaurantId: 0, clientId: 0, deliverymanId: 0 };
+      return {};
+  }
+}
+function filterQueryToRole(userId, roleLabel, query) {
+  let filter = {};
+  switch (roleLabel) {
+    case "user":
+      return { clientId: { $eq: userId } };
+    case "deleveryman":
+      return { deliverymanId: { $eq: userId } };
+    case "restaurantOwner":
+      return { restaurantId: { $eq: userId } };
+    default:
+      if (query.restaurantid) filter["restaurantId"] = { $eq: query.restaurantid }
+      if (query.deliverymanid) filter["deliverymanId"] = { $eq: query.deliverymanid }
+      if (query.clienttid) filter["clientId"] = { $eq: query.clienttid }
+      return filter
   }
 }
 module.exports = {
   getOrderbyId: async (req, res) => {
     const { id } = req.params;
+    const { userId, roleLabel } = req.query;
+
+    const format = formatResponseToRole(roleLabel);
+    const filter = filterQueryToRole(userId, roleLabel, { restaurantid, clienttid, deliverymanid });
+    filter["id"] = id;
 
     if (!isValidObjectId(id)) return errors.invalidId;
-    const { userId, roleLabel } = req.query;
-    const format = formatResponseToRole(roleLabel);
 
-    const orders = await Order.findById(id, format);
-    return orders;
-  },
-  getOrderbyClientId: async (req, res) => {
-    const { id } = req.params;
-
-    if (!isValidObjectId(id)) return errors.invalidId;
-    const { userId, roleLabel } = req.query;
-    const format = formatResponseToRole(roleLabel);
-
-    const orders = await Order.find({ clientId: { $eq: id } }, format);
-    return orders;
-  },
-  getOrderbyRestaurantId: async (req, res) => {
-    const { id } = req.params;
-
-    if (!isValidObjectId(id)) return errors.invalidId;
-    const { userId, roleLabel } = req.query;
-    const format = formatResponseToRole(roleLabel);
-
-    const orders = await Order.find({ restaurantId: { $eq: id } }, format);
-    return orders;
-  },
-  getOrderbyDeliverymanId: async (req, res) => {
-    const { id } = req.params;
-
-    if (!isValidObjectId(id)) return errors.invalidId;
-    const { userId, roleLabel } = req.query;
-    const format = formatResponseToRole(roleLabel);
-
-    const orders = await Order.find({ deliverymanId: { $eq: id } }, format);
+    const orders = await Order.findById(filter, format);
     return orders;
   },
   getOrders: async (req, res) => {
-    const { userId, roleLabel } = req.query;
+    const { userId, roleLabel, restaurantid, clienttid, deliverymanid } = req.query;
+
     const format = formatResponseToRole(roleLabel);
-    const allOrders = await Order.find({}, format);
+    const filter = filterQueryToRole(userId, roleLabel, { restaurantid, clienttid, deliverymanid });
+
+    const allOrders = await Order.find(filter, format);
     return allOrders;
   },
   patchOrder: async (req, res) => {
     const { id } = req.params;
     const { articleIdList, date, clientCode, status, restaurantId, clientId, deliverymanId } = req.body;
+    const { userId, roleLabel } = req.query;
+
     const validatedDate = new Date(date);
     const statusId = await Status.findOne({ state: { $eq: status } });
+    const targetOrder = await Order.findById(id);
 
+    if (roleLabel == "user" && targetOrder.clientId != userId) return errors.invalidPermissions;
     if (!isValidObjectId(id)) return errors.invalidId;
     if (!articleIdList && !date && !clientCode && !status && !restaurantId && !clientId && !deliverymanId) return errors.missingRequiredParams;
     if (date && (!validatedDate || validatedDate == "Invalid Date")) return errors.invalidDateFormat;
@@ -110,9 +110,13 @@ module.exports = {
   putOrder: async (req, res) => {
     const { id } = req.params;
     const { articleIdList, date, clientCode, status, restaurantId, clientId, deliverymanId } = req.body;
+    const { userId, roleLabel } = req.query;
+
     const validatedDate = new Date(date);
     const statusId = await Status.findOne({ state: { $eq: status } });
+    const targetOrder = await Order.findById(id);
 
+    if (roleLabel == "user" && targetOrder.clientId != userId) return errors.invalidPermissions;
     if (!isValidObjectId(id)) return errors.invalidId;
     if (!articleIdList || !date || !clientCode || !status || !restaurantId || !clientId || !deliverymanId) return errors.missingRequiredParams;
     if (date && (!validatedDate || validatedDate == "Invalid Date")) return errors.invalidDateFormat;
@@ -123,11 +127,13 @@ module.exports = {
   },
   deleteOrder: async (req, res) => {
     const { id } = req.params;
+    const { userId, roleLabel } = req.query;
+    const targetOrder = await Order.findById(id);
 
+    if (roleLabel == "user" && targetOrder.clientId != userId) return errors.invalidPermissions;
     if (!isValidObjectId(id)) return errors.invalidId;
 
     const a = await Order.findByIdAndDelete(id)
-    console.log(a);
     return 'Order deleted successfully';
   },
   createOrder: async (req, res) => {
@@ -139,7 +145,7 @@ module.exports = {
     if (!date || !clientCode || !restaurantId || !clientId) return errors.missingRequiredParams;
     if (date && (!validatedDate || validatedDate == "Invalid Date")) return errors.invalidDateFormat;
     if (status && !statusId) return errors.statusNotFound;
-    
+
     await Order.create({ articleIdList, date, clientCode, statusId, restaurantId, clientId, deliverymanId });
     return 'Order created successfully';
   }
